// src/types/theme.types.ts
export type CustomButtonProps = {
  variant: 'text' | 'outlined' | 'contained' | 'light' | undefined;
  color:
    | 'inherit'
    | 'primary'
    | 'secondary'
    | 'secondary2'
    | 'secondary3'
    | 'success'
    | 'error'
    | 'info'
    | 'warning'
    | undefined;
  size: 'small' | 'large' | 'medium' | undefined;
};